+++ Packages/com.fang.framework/Tests/Runtime/ScopeLifecycleTests.cs
@@ -10,54 +10,72 @@ namespace Fang.Framework.Tests
             LifecycleLog.Reset();
         }
 
+        [TearDown]
+        public void TearDown()
+        {
+            ProbeScope.DestroyAll();
+        }
+
         [Test]
         public void AddService_creates_then_initializes_in_add_order()
         {
-            var scope = new ProbeScope();
+            var scope = ProbeScope.Create<ProbeScope>();
 
             scope.Add<LifecycleProbeA>();
             scope.Add<LifecycleProbeB>();
             scope.Add<LifecycleProbeC>();
 
             CollectionAssert.AreEqual(
-                new[] { "create:A", "init:A", "create:B", "init:B", "create:C", "init:C" },
+                new[] { "init:A", "init:B", "init:C" },
                 LifecycleLog.Entries);
         }
 
         [Test]
         public void Dispose_disposes_services_in_reverse_add_order_exactly_once()
         {
-            var scope = new ProbeScope();
+            var scope = ProbeScope.Create<ProbeScope>();
             scope.Add<LifecycleProbeA>();
             scope.Add<LifecycleProbeB>();
 
-            scope.Dispose();
-            scope.Dispose();
+            scope.OnDispose();
+            scope.OnDispose();
 
             CollectionAssert.AreEqual(
-                new[] { "create:A", "init:A", "create:B", "init:B", "dispose:B", "dispose:A" },
+                new[] { "init:A", "init:B", "dispose:B", "dispose:A" },
                 LifecycleLog.Entries);
+        }
+
+        [Test]
+        public void OnDispose_is_a_no_op_before_OnInit()
+        {
+            var scope = ProbeScope.CreateUninitialized<ProbeScope>();
+            scope.Add<AudioService>();
+
+            scope.OnDispose();
+
+            Assert.IsFalse(scope.IsInitialized);
+            Assert.AreEqual(1, scope.Services.Count);
         }
 
         [Test]
         public void Dispose_disposes_children_before_own_services()
         {
-            var root = new ProbeScope();
+            var root = ProbeScope.Create<ProbeScope>();
             root.Add<LifecycleProbeA>();
             var child = root.AddChild<ProbeChildScope>();
             child.Add<LifecycleProbeB>();
 
-            root.Dispose();
+            root.OnDispose();
 
             CollectionAssert.AreEqual(
-                new[] { "create:A", "init:A", "create:B", "init:B", "dispose:B", "dispose:A" },
+                new[] { "init:A", "init:B", "dispose:B", "dispose:A" },
                 LifecycleLog.Entries);
         }
 
         [Test]
         public void Tick_drives_own_services_in_add_order_then_children()
         {
-            var root = new ProbeScope();
+            var root = ProbeScope.Create<ProbeScope>();
             root.Add<TickProbeService>();
             root.Add<FixedOnlyProbeService>();
             root.Add<SecondaryTickProbeService>();
@@ -72,7 +90,7 @@ namespace Fang.Framework.Tests
         [Test]
         public void FixedTick_drives_own_services_in_add_order_then_children()
         {
-            var root = new ProbeScope();
+            var root = ProbeScope.Create<ProbeScope>();
             root.Add<FixedOnlyProbeService>();
             root.Add<TickProbeService>();
             root.Add<SecondaryTickProbeService>();
@@ -89,7 +107,7 @@ namespace Fang.Framework.Tests
         [Test]
         public void Tick_forwards_the_delta_time()
         {
-            var scope = new ProbeScope();
+            var scope = ProbeScope.Create<ProbeScope>();
             var service = scope.Add<DeltaProbeService>();
 
             scope.Tick(0.25f);
@@ -102,22 +120,22 @@ namespace Fang.Framework.Tests
         [Test]
         public void Tick_skips_services_without_the_capability()
         {
-            var scope = new ProbeScope();
+            var scope = ProbeScope.Create<ProbeScope>();
             scope.Add<LifecycleProbeA>();
 
             scope.Tick(1f);
             scope.FixedTick(1f);
 
-            CollectionAssert.AreEqual(new[] { "create:A", "init:A" }, LifecycleLog.Entries);
+            CollectionAssert.AreEqual(new[] { "init:A" }, LifecycleLog.Entries);
         }
 
         [Test]
         public void Tick_does_not_touch_a_disposed_child()
         {
-            var root = new ProbeScope();
+            var root = ProbeScope.Create<ProbeScope>();
             var child = root.AddChild<ProbeChildScope>();
             child.Add<TickProbeService>();
-            child.Dispose();
+            child.OnDispose();
 
             root.Tick(1f);