+++ Packages/com.fang.framework/Tests/Runtime/ScopeRegistrationTests.cs
@@ -11,10 +11,16 @@ namespace Fang.Framework.Tests
             LifecycleLog.Reset();
         }
 
+        [TearDown]
+        public void TearDown()
+        {
+            ProbeScope.DestroyAll();
+        }
+
         [Test]
         public void AddService_injects_the_owning_scope_and_returns_the_instance()
         {
-            var scope = new ProbeScope();
+            var scope = ProbeScope.Create<ProbeScope>();
 
             var service = scope.Add<AudioService>();
 
@@ -26,17 +32,40 @@ namespace Fang.Framework.Tests
         [Test]
         public void AddService_initializes_the_service_right_away()
         {
-            var scope = new ProbeScope();
+            var scope = ProbeScope.Create<ProbeScope>();
 
             scope.Add<LifecycleProbeA>();
 
-            CollectionAssert.AreEqual(new[] { "create:A", "init:A" }, LifecycleLog.Entries);
+            CollectionAssert.AreEqual(new[] { "init:A" }, LifecycleLog.Entries);
+        }
+
+        [Test]
+        public void AddService_places_the_service_under_the_services_node()
+        {
+            var scope = ProbeScope.Create<ProbeScope>();
+
+            var service = scope.Add<AudioService>();
+
+            Assert.AreEqual("Services", service.transform.parent.name);
+            Assert.AreSame(scope.transform, service.transform.parent.parent);
+        }
+
+        [Test]
+        public void OnInit_flips_IsInitialized()
+        {
+            var scope = ProbeScope.CreateUninitialized<ProbeScope>();
+
+            Assert.IsFalse(scope.IsInitialized);
+
+            scope.OnInit();
+
+            Assert.IsTrue(scope.IsInitialized);
         }
 
         [Test]
         public void GetService_throws_for_an_unregistered_service()
         {
-            var scope = new ProbeScope();
+            var scope = ProbeScope.Create<ProbeScope>();
 
             Assert.Throws<InvalidOperationException>(() => { scope.GetService<AudioService>(); });
         }
@@ -44,7 +73,7 @@ namespace Fang.Framework.Tests
         [Test]
         public void GetService_message_names_the_service_and_the_scope()
         {
-            var scope = new ProbeScope();
+            var scope = ProbeScope.Create<ProbeScope>();
 
             var exception = Assert.Throws<InvalidOperationException>(() => { scope.GetService<AudioService>(); });
 
@@ -55,20 +84,32 @@ namespace Fang.Framework.Tests
         [Test]
         public void RemoveService_disposes_and_unregisters_the_service()
         {
-            var scope = new ProbeScope();
+            var scope = ProbeScope.Create<ProbeScope>();
             scope.Add<LifecycleProbeA>();
 
             scope.Remove<LifecycleProbeA>();
 
-            CollectionAssert.AreEqual(new[] { "create:A", "init:A", "dispose:A" }, LifecycleLog.Entries);
+            CollectionAssert.AreEqual(new[] { "init:A", "dispose:A" }, LifecycleLog.Entries);
             Assert.AreEqual(0, scope.Services.Count);
             Assert.Throws<InvalidOperationException>(() => { scope.GetService<LifecycleProbeA>(); });
         }
 
         [Test]
+        public void RemoveService_keeps_the_game_object_alive()
+        {
+            var scope = ProbeScope.Create<ProbeScope>();
+            var service = scope.Add<AudioService>();
+
+            scope.Remove<AudioService>();
+
+            Assert.IsTrue(service != null);
+            Assert.Throws<InvalidOperationException>(() => { scope.GetService<AudioService>(); });
+        }
+
+        [Test]
         public void RemoveService_is_a_no_op_when_the_service_is_absent()
         {
-            var scope = new ProbeScope();
+            var scope = ProbeScope.Create<ProbeScope>();
 
             Assert.DoesNotThrow(() => { scope.Remove<AudioService>(); });
         }
@@ -76,7 +117,7 @@ namespace Fang.Framework.Tests
         [Test]
         public void GetService_returns_the_first_match_in_add_order()
         {
-            var scope = new ProbeScope();
+            var scope = ProbeScope.Create<ProbeScope>();
             var first = scope.Add<AudioService>();
             scope.Add<AudioServiceOverride>();